Add example tf project and update compiler entry
We don't need to support multiple languages or formats right now
but I'm not sure how we handle terraform from sources like cdktf
yet. Or if we want to also add in cloudformation parsing later for
a giga tool that processes and generates output.

Compiler works as expected and simply grabs hcl .tf files to parse
for now with required defaults but can be extended later easily.