Extend the tree-sitter-hcl parser so we can name nodes
Using the hcl parser standalone works but means we have to write a
lot of queries. If we can instead name the nodes better and avoid
queries in the compiler it could allow for an interface that is
capable of loading *any* hcl (not just terraform) into adk and
outputting events.