Fix build location of the language wasm
When running webpack the location of the language wasm
is different than the location during dev. To allow for using ts-node,
the tree sitter playground, and webpack together easily and not
copying the wasm around everywhere I added a flag to signify that
the location is different when built with webpack.